# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Chronicle API
  version: 1.0.0
  description: REST API for Chronicle, a self-hosted TTRPG worldbuilding platform. Provides programmatic access to campaign data for external tool integration (Foundry VTT, custom scripts).
servers:
  - url: "{baseUrl}/api/v1"
    variables:
      baseUrl:
        default: http://localhost:8080
        description: Chronicle server base URL
security:
  - BearerApiKey: []
...
paths:
...
  /campaigns/{campaignId}:
    get:
      tags: [Campaigns]
      summary: Get campaign details
      operationId: getCampaign
      parameters:
        - $ref: "#/components/parameters/campaignId"
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # --- Entity Types ---
  /campaigns/{campaignId}/entity-types:
    get:
      tags: [Entities]
      summary: List entity types
      operationId: listEntityTypes
      parameters:
        - $ref: "#/components/parameters/campaignId"
      responses:
        "200":
          description: All entity types for the campaign
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EntityType"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /campaigns/{campaignId}/entity-types/{typeId}:
    get:
      tags: [Entities]
      summary: Get entity type
      operationId: getEntityType
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/typeId"
      responses:
        "200":
          description: Entity type details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityType"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

  # --- Entities ---
  /campaigns/{campaignId}/entities:
    get:
      tags: [Entities]
      summary: List entities
      operationId: listEntities
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: type_id
          in: query
          schema:
            type: integer
        - name: q
          in: query
          description: Search query
          schema:
            type: string
      responses:
        "200":
          description: Paginated entity list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedEntities"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Entities]
      summary: Create entity
      operationId: createEntity
      parameters:
        - $ref: "#/components/parameters/campaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntityRequest"
      responses:
        "201":
          description: Created entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/entities/{entityId}:
    get:
      tags: [Entities]
      summary: Get entity
      operationId: getEntity
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/entityId"
      responses:
        "200":
          description: Entity details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    put:
      tags: [Entities]
      summary: Update entity
      operationId: updateEntity
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/entityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntityRequest"
      responses:
        "200":
          description: Updated entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      tags: [Entities]
      summary: Delete entity
      operationId: deleteEntity
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/entityId"
      responses:
        "204":
          description: Entity deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /campaigns/{campaignId}/entities/{entityId}/fields:
    put:
      tags: [Entities]
      summary: Update entity fields
      operationId: updateEntityFields
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/entityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fields_data:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Fields updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /campaigns/{campaignId}/entities/{entityId}/relations:
    get:
      tags: [Entities]
      summary: List entity relations
      operationId: listEntityRelations
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - $ref: "#/components/parameters/entityId"
      responses:
        "200":
          description: List of relations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Relation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # --- Calendar ---
...
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: >
        API key passed as Bearer token. Keys are scoped to a single campaign
        with permission levels: read, write, sync. Rate limited per key.
  parameters:
    campaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
    entityId:
      name: entityId
      in: path
      required: true
      schema:
        type: string
    typeId:
      name: typeId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
        type:
          type: string
        message:
          type: string
    StatusOk:
      type: object
      properties:
        status:
          type: string
          example: ok

    # --- Campaigns ---
    Campaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        is_public:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Entities ---
    EntityType:
      type: object
      properties:
        id:
          type: integer
        campaign_id:
          type: string
        slug:
          type: string
        name:
          type: string
        name_plural:
          type: string
        icon:
          type: string
        color:
          type: string
        description:
          type: string
          nullable: true
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FieldDefinition"
        layout:
          type: object
          description: Row-based page layout
        sort_order:
          type: integer
        is_default:
          type: boolean
        enabled:
          type: boolean
    FieldDefinition:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, number, select, textarea, checkbox, url]
        section:
          type: string
        options:
          type: array
          items:
            type: string
    Entity:
      type: object
      properties:
        id:
          type: string
        campaign_id:
          type: string
        entity_type_id:
          type: integer
        name:
          type: string
        slug:
          type: string
        entry:
          type: string
          nullable: true
          description: ProseMirror JSON document
        entry_html:
          type: string
          nullable: true
        image_path:
          type: string
          nullable: true
        parent_id:
          type: string
          nullable: true
        type_label:
          type: string
          nullable: true
        is_private:
          type: boolean
        is_template:
          type: boolean
        fields_data:
          type: object
          additionalProperties: true
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        type_name:
          type: string
        type_icon:
          type: string
        type_color:
          type: string
        type_slug:
          type: string
        tags:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              color:
                type: string
    PaginatedEntities:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
    CreateEntityRequest:
      type: object
      required: [name, entity_type_id]
      properties:
        name:
          type: string
        entity_type_id:
          type: integer
        type_label:
          type: string
        is_private:
          type: boolean
          default: false
        fields_data:
          type: object
          additionalProperties: true
    UpdateEntityRequest:
      type: object
      properties:
        name:
          type: string
        type_label:
          type: string
        is_private:
          type: boolean
        entry:
          type: string
          description: ProseMirror JSON document
        fields_data:
          type: object
          additionalProperties: true
    Relation:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: string
        sourceEntityId:
          type: string
        targetEntityId:
          type: string
        relationType:
          type: string
        reverseRelationType:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        targetEntityName:
          type: string
        targetEntityIcon:
          type: string
        targetEntityColor:
          type: string
        targetEntitySlug:
          type: string
        targetEntityType:
          type: string

    # --- Calendar ---
...